+++ src/components/styles/Loading.css
@@ -1,67 +1,68 @@
-.loader,
-.loader:before,
-.loader:after {
-  border-radius: 50%;
-}
 .loader {
-  color: #7b7fad;
-  font-size: 11px;
-  text-indent: -99999em;
-  margin: 55px auto;
-  position: relative;
-  width: 10em;
-  height: 10em;
-  box-shadow: inset 0 0 0 1em;
-  -webkit-transform: translateZ(0);
-  -ms-transform: translateZ(0);
-  transform: translateZ(0);
-}
-.loader:before,
-.loader:after {
-  position: absolute;
-  content: '';
-}
-.loader:before {
-  width: 5.2em;
-  height: 10.2em;
-  background: #eaecf0;
-  border-radius: 10.2em 0 0 10.2em;
-  top: -0.1em;
-  left: -0.1em;
-  -webkit-transform-origin: 5.1em 5.1em;
-  transform-origin: 5.1em 5.1em;
-  -webkit-animation: load2 2s infinite ease 1.5s;
-  animation: load2 2s infinite ease 1.5s;
-}
-.loader:after {
-  width: 5.2em;
-  height: 10.2em;
-  background: #eaecf0;
-  border-radius: 0 10.2em 10.2em 0;
-  top: -0.1em;
-  left: 4.9em;
-  -webkit-transform-origin: 0.1em 5.1em;
-  transform-origin: 0.1em 5.1em;
-  -webkit-animation: load2 2s infinite ease;
-  animation: load2 2s infinite ease;
-}
-@-webkit-keyframes load2 {
-  0% {
-    -webkit-transform: rotate(0deg);
-    transform: rotate(0deg);
+    margin: 100px auto;
+    font-size: 25px;
+    width: 1em;
+    height: 1em;
+    border-radius: 50%;
+    position: relative;
+    text-indent: -9999em;
+    -webkit-animation: load5 1.1s infinite ease;
+    animation: load5 1.1s infinite ease;
+    -webkit-transform: translateZ(0);
+    -ms-transform: translateZ(0);
+    transform: translateZ(0);
   }
-  100% {
-    -webkit-transform: rotate(360deg);
-    transform: rotate(360deg);
+  @-webkit-keyframes load5 {
+    0%,
+    100% {
+      box-shadow: 0em -2.6em 0em 0em #43404f, 1.8em -1.8em 0 0em rgba(67,64,79, 0.2), 2.5em 0em 0 0em rgba(67,64,79, 0.2), 1.75em 1.75em 0 0em rgba(67,64,79, 0.2), 0em 2.5em 0 0em rgba(67,64,79, 0.2), -1.8em 1.8em 0 0em rgba(67,64,79, 0.2), -2.6em 0em 0 0em rgba(67,64,79, 0.5), -1.8em -1.8em 0 0em rgba(67,64,79, 0.7);
+    }
+    12.5% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.7), 1.8em -1.8em 0 0em #43404f, 2.5em 0em 0 0em rgba(67,64,79, 0.2), 1.75em 1.75em 0 0em rgba(67,64,79, 0.2), 0em 2.5em 0 0em rgba(67,64,79, 0.2), -1.8em 1.8em 0 0em rgba(67,64,79, 0.2), -2.6em 0em 0 0em rgba(67,64,79, 0.2), -1.8em -1.8em 0 0em rgba(67,64,79, 0.5);
+    }
+    25% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.5), 1.8em -1.8em 0 0em rgba(67,64,79, 0.7), 2.5em 0em 0 0em #43404f, 1.75em 1.75em 0 0em rgba(67,64,79, 0.2), 0em 2.5em 0 0em rgba(67,64,79, 0.2), -1.8em 1.8em 0 0em rgba(67,64,79, 0.2), -2.6em 0em 0 0em rgba(67,64,79, 0.2), -1.8em -1.8em 0 0em rgba(67,64,79, 0.2);
+    }
+    37.5% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.2), 1.8em -1.8em 0 0em rgba(67,64,79, 0.5), 2.5em 0em 0 0em rgba(67,64,79, 0.7), 1.75em 1.75em 0 0em #43404f, 0em 2.5em 0 0em rgba(67,64,79, 0.2), -1.8em 1.8em 0 0em rgba(67,64,79, 0.2), -2.6em 0em 0 0em rgba(67,64,79, 0.2), -1.8em -1.8em 0 0em rgba(67,64,79, 0.2);
+    }
+    50% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.2), 1.8em -1.8em 0 0em rgba(67,64,79, 0.2), 2.5em 0em 0 0em rgba(67,64,79, 0.5), 1.75em 1.75em 0 0em rgba(67,64,79, 0.7), 0em 2.5em 0 0em #43404f, -1.8em 1.8em 0 0em rgba(67,64,79, 0.2), -2.6em 0em 0 0em rgba(67,64,79, 0.2), -1.8em -1.8em 0 0em rgba(67,64,79, 0.2);
+    }
+    62.5% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.2), 1.8em -1.8em 0 0em rgba(67,64,79, 0.2), 2.5em 0em 0 0em rgba(67,64,79, 0.2), 1.75em 1.75em 0 0em rgba(67,64,79, 0.5), 0em 2.5em 0 0em rgba(67,64,79, 0.7), -1.8em 1.8em 0 0em #43404f, -2.6em 0em 0 0em rgba(67,64,79, 0.2), -1.8em -1.8em 0 0em rgba(67,64,79, 0.2);
+    }
+    75% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.2), 1.8em -1.8em 0 0em rgba(67,64,79, 0.2), 2.5em 0em 0 0em rgba(67,64,79, 0.2), 1.75em 1.75em 0 0em rgba(67,64,79, 0.2), 0em 2.5em 0 0em rgba(67,64,79, 0.5), -1.8em 1.8em 0 0em rgba(67,64,79, 0.7), -2.6em 0em 0 0em #43404f, -1.8em -1.8em 0 0em rgba(67,64,79, 0.2);
+    }
+    87.5% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.2), 1.8em -1.8em 0 0em rgba(67,64,79, 0.2), 2.5em 0em 0 0em rgba(67,64,79, 0.2), 1.75em 1.75em 0 0em rgba(67,64,79, 0.2), 0em 2.5em 0 0em rgba(67,64,79, 0.2), -1.8em 1.8em 0 0em rgba(67,64,79, 0.5), -2.6em 0em 0 0em rgba(67,64,79, 0.7), -1.8em -1.8em 0 0em #43404f;
+    }
   }
-}
-@keyframes load2 {
-  0% {
-    -webkit-transform: rotate(0deg);
-    transform: rotate(0deg);
+  @keyframes load5 {
+    0%,
+    100% {
+      box-shadow: 0em -2.6em 0em 0em #43404f, 1.8em -1.8em 0 0em rgba(67,64,79, 0.2), 2.5em 0em 0 0em rgba(67,64,79, 0.2), 1.75em 1.75em 0 0em rgba(67,64,79, 0.2), 0em 2.5em 0 0em rgba(67,64,79, 0.2), -1.8em 1.8em 0 0em rgba(67,64,79, 0.2), -2.6em 0em 0 0em rgba(67,64,79, 0.5), -1.8em -1.8em 0 0em rgba(67,64,79, 0.7);
+    }
+    12.5% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.7), 1.8em -1.8em 0 0em #43404f, 2.5em 0em 0 0em rgba(67,64,79, 0.2), 1.75em 1.75em 0 0em rgba(67,64,79, 0.2), 0em 2.5em 0 0em rgba(67,64,79, 0.2), -1.8em 1.8em 0 0em rgba(67,64,79, 0.2), -2.6em 0em 0 0em rgba(67,64,79, 0.2), -1.8em -1.8em 0 0em rgba(67,64,79, 0.5);
+    }
+    25% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.5), 1.8em -1.8em 0 0em rgba(67,64,79, 0.7), 2.5em 0em 0 0em #43404f, 1.75em 1.75em 0 0em rgba(67,64,79, 0.2), 0em 2.5em 0 0em rgba(67,64,79, 0.2), -1.8em 1.8em 0 0em rgba(67,64,79, 0.2), -2.6em 0em 0 0em rgba(67,64,79, 0.2), -1.8em -1.8em 0 0em rgba(67,64,79, 0.2);
+    }
+    37.5% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.2), 1.8em -1.8em 0 0em rgba(67,64,79, 0.5), 2.5em 0em 0 0em rgba(67,64,79, 0.7), 1.75em 1.75em 0 0em #43404f, 0em 2.5em 0 0em rgba(67,64,79, 0.2), -1.8em 1.8em 0 0em rgba(67,64,79, 0.2), -2.6em 0em 0 0em rgba(67,64,79, 0.2), -1.8em -1.8em 0 0em rgba(67,64,79, 0.2);
+    }
+    50% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.2), 1.8em -1.8em 0 0em rgba(67,64,79, 0.2), 2.5em 0em 0 0em rgba(67,64,79, 0.5), 1.75em 1.75em 0 0em rgba(67,64,79, 0.7), 0em 2.5em 0 0em #43404f, -1.8em 1.8em 0 0em rgba(67,64,79, 0.2), -2.6em 0em 0 0em rgba(67,64,79, 0.2), -1.8em -1.8em 0 0em rgba(67,64,79, 0.2);
+    }
+    62.5% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.2), 1.8em -1.8em 0 0em rgba(67,64,79, 0.2), 2.5em 0em 0 0em rgba(67,64,79, 0.2), 1.75em 1.75em 0 0em rgba(67,64,79, 0.5), 0em 2.5em 0 0em rgba(67,64,79, 0.7), -1.8em 1.8em 0 0em #43404f, -2.6em 0em 0 0em rgba(67,64,79, 0.2), -1.8em -1.8em 0 0em rgba(67,64,79, 0.2);
+    }
+    75% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.2), 1.8em -1.8em 0 0em rgba(67,64,79, 0.2), 2.5em 0em 0 0em rgba(67,64,79, 0.2), 1.75em 1.75em 0 0em rgba(67,64,79, 0.2), 0em 2.5em 0 0em rgba(67,64,79, 0.5), -1.8em 1.8em 0 0em rgba(67,64,79, 0.7), -2.6em 0em 0 0em #43404f, -1.8em -1.8em 0 0em rgba(67,64,79, 0.2);
+    }
+    87.5% {
+      box-shadow: 0em -2.6em 0em 0em rgba(67,64,79, 0.2), 1.8em -1.8em 0 0em rgba(67,64,79, 0.2), 2.5em 0em 0 0em rgba(67,64,79, 0.2), 1.75em 1.75em 0 0em rgba(67,64,79, 0.2), 0em 2.5em 0 0em rgba(67,64,79, 0.2), -1.8em 1.8em 0 0em rgba(67,64,79, 0.5), -2.6em 0em 0 0em rgba(67,64,79, 0.7), -1.8em -1.8em 0 0em #43404f;
+    }
   }
-  100% {
-    -webkit-transform: rotate(360deg);
-    transform: rotate(360deg);
-  }
-}
